requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
